chore: address minor review findings from the add-cap branch
- Widen the set-instance-variable cheat's value from int32 to int64, matching
  WorldInstance's storage. A wire change, so game::ProtocolVersion goes to 10
  and the manifest and docs/protocol_versions.md record it.
- Include the two new AI headers by their full path, as everything else in
  game_server does.
- Document the two new cheat declarations.
- Have crypt_add_cap.lua put the character back on the default map and spawn
  point and drop godmode. Its arrival assertion was tautological on a re-run
  that started already standing in the crypt, and a character left inside a
  dungeon instance strands every scenario that moves relative to where it is.

+++ src/e2e_client/bot_realm_connector.h
@@ -337,7 +337,12 @@ namespace mmo
 		/// GAME MASTER only. Toggles damage immunity on the sender's character.
 		void CheatGodmode(bool enable);
 
-		void CheatSetInstanceVariable(uint32 key, int32 value);
+		/// GAME MASTER only. Sets an instance-scoped variable in the world instance the character is
+		///	in, so a scenario can put an encounter into a given state without having to reach it
+		///	through gameplay.
+		///	@param key The variable key, matching the one an InstanceVariable condition reads.
+		///	@param value The value to assign.
+		void CheatSetInstanceVariable(uint32 key, int64 value);
 
 		/// GAME MASTER only. Accepts the given quest without a questgiver interaction.
 		void CheatAcceptQuest(uint32 questId);
